Split into top level header + sub headers
